Fix clippy warnings
