+++ Cargo.toml
@@ -21,6 +21,5 @@ codegen-units = 1
 [patch.crates-io]
 zcash_encoding = { path = "components/zcash_encoding" }
 zcash_note_encryption = { path = "components/zcash_note_encryption" }
-orchard = { version = "0.4", git = "https://github.com/QED-it/orchard", branch = "librustzcash_980736806" }
-bridgetree = { git = "https://github.com/zcash/incrementalmerkletree.git", rev = "ea1686e8f8f6c1e41aa97251a7eb4fadfd33df47" }
-incrementalmerkletree = { git = "https://github.com/zcash/incrementalmerkletree.git", rev = "ea1686e8f8f6c1e41aa97251a7eb4fadfd33df47" }
+#orchard = { version = "0.5", git = "https://github.com/QED-it/orchard", branch = "upgrade_librustzcash_for_orchard_v05" }
+orchard = { version = "0.5", path = "../orchard" }
