+++ components/zcash_address/CHANGELOG.md
@@ -7,6 +7,10 @@ and this library adheres to Rust's notion of
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-06-06
+### Changed
+- Bumped bs58 dependency to `0.5`.
+
 ## [0.2.1] - 2023-04-15
 ### Changed
 - Bumped internal dependency to `bech32 0.9`.
