+++ zcash_client_backend/Cargo.toml
@@ -20,10 +20,11 @@ exclude = ["*.proto"]
 development = ["zcash_proofs"]
 
 [dependencies]
-zcash_address = { version = "0.2", path = "../components/zcash_address" }
+incrementalmerkletree = { version = "0.4", features = ["legacy-api"] }
+zcash_address = { version = "0.3", path = "../components/zcash_address" }
 zcash_encoding = { version = "0.2", path = "../components/zcash_encoding" }
-zcash_note_encryption = "0.3"
-zcash_primitives = { version = "0.11", path = "../zcash_primitives", default-features = false }
+zcash_note_encryption = "0.4"
+zcash_primitives = { version = "0.12", path = "../zcash_primitives", default-features = false }
 
 # Dependencies exposed in a public API:
 # (Breaking upgrades to these require a breaking upgrade to this crate.)
@@ -33,7 +34,7 @@ time = "0.2"
 # - Encodings
 base64 = "0.21"
 bech32 = "0.9"
-bs58 = { version = "0.4", features = ["check"] }
+bs58 = { version = "0.5", features = ["check"] }
 
 # - Errors
 hdwallet = { version = "0.4", optional = true }
@@ -53,7 +54,7 @@ subtle = "2.2.3"
 # - Shielded protocols
 bls12_381 = "0.8"
 group = "0.13"
-orchard = { version = "0.4", default-features = false }
+orchard = { version = "0.5", default-features = false }
 
 # - Test dependencies
 proptest = { version = "1.0.0", optional = true }
@@ -83,9 +84,9 @@ jubjub = "0.10"
 proptest = "1.0.0"
 rand_core = "0.6"
 rand_xorshift = "0.3"
-tempfile = "3.1.0"
-zcash_proofs = { version = "0.11", path = "../zcash_proofs", default-features = false }
-zcash_address = { version = "0.2", path = "../components/zcash_address", features = ["test-dependencies"] }
+tempfile = "3.5.0"
+zcash_proofs = { version = "0.12", path = "../zcash_proofs", default-features = false }
+zcash_address = { version = "0.3", path = "../components/zcash_address", features = ["test-dependencies"] }
 
 [features]
 lightwalletd-tonic = ["tonic"]
@@ -94,6 +95,7 @@ test-dependencies = [
     "proptest",
     "orchard/test-dependencies",
     "zcash_primitives/test-dependencies",
+    "incrementalmerkletree/test-dependencies"
 ]
 unstable = ["byteorder"]
