+++ zcash_client_sqlite/src/wallet.rs
@@ -64,7 +64,6 @@
 //!   wallet.
 //! - `memo` the shielded memo associated with the output, if any.
 
-use group::ff::PrimeField;
 use rusqlite::{named_params, OptionalExtension, ToSql};
 use std::collections::HashMap;
 use std::convert::TryFrom;
@@ -73,8 +72,7 @@ use zcash_primitives::{
     block::BlockHash,
     consensus::{self, BlockHeight, BranchId, NetworkUpgrade, Parameters},
     memo::{Memo, MemoBytes},
-    merkle_tree::{read_commitment_tree, read_incremental_witness},
-    sapling::{self, Note, Nullifier},
+    sapling::CommitmentTree,
     transaction::{components::Amount, Transaction, TxId},
     zip32::{
         sapling::{DiversifiableFullViewingKey, ExtendedFullViewingKey},
@@ -86,13 +84,11 @@ use zcash_client_backend::{
     address::{RecipientAddress, UnifiedAddress},
     data_api::{PoolType, Recipient, SentTransactionOutput},
     keys::UnifiedFullViewingKey,
-    wallet::{WalletSaplingOutput, WalletTx},
-    DecryptedOutput, TransferType,
+    wallet::WalletTx,
 };
 
 use crate::{
-    error::SqliteClientError, prepared::InsertAddress, DataConnStmtCache, NoteId, WalletDb,
-    PRUNING_HEIGHT,
+    error::SqliteClientError, prepared::InsertAddress, DataConnStmtCache, WalletDb, PRUNING_HEIGHT,
 };
 
 #[cfg(feature = "transparent-inputs")]
@@ -110,7 +106,7 @@ use {
 };
 
 pub mod init;
-pub mod transact;
+pub(crate) mod sapling;
 
 pub(crate) fn pool_code(pool_type: PoolType) -> i64 {
     // These constants are *incidentally* shared with the typecodes
@@ -119,59 +115,6 @@ pub(crate) fn pool_code(pool_type: PoolType) -> i64 {
     match pool_type {
         PoolType::Transparent => 0i64,
         PoolType::Sapling => 2i64,
-    }
-}
-
-/// This trait provides a generalization over shielded output representations.
-pub(crate) trait ReceivedSaplingOutput {
-    fn index(&self) -> usize;
-    fn account(&self) -> AccountId;
-    fn note(&self) -> &Note;
-    fn memo(&self) -> Option<&MemoBytes>;
-    fn is_change(&self) -> bool;
-    fn nullifier(&self) -> Option<&Nullifier>;
-}
-
-impl ReceivedSaplingOutput for WalletSaplingOutput<Nullifier> {
-    fn index(&self) -> usize {
-        self.index()
-    }
-    fn account(&self) -> AccountId {
-        WalletSaplingOutput::account(self)
-    }
-    fn note(&self) -> &Note {
-        WalletSaplingOutput::note(self)
-    }
-    fn memo(&self) -> Option<&MemoBytes> {
-        None
-    }
-    fn is_change(&self) -> bool {
-        WalletSaplingOutput::is_change(self)
-    }
-
-    fn nullifier(&self) -> Option<&Nullifier> {
-        Some(self.nf())
-    }
-}
-
-impl ReceivedSaplingOutput for DecryptedOutput<Note> {
-    fn index(&self) -> usize {
-        self.index
-    }
-    fn account(&self) -> AccountId {
-        self.account
-    }
-    fn note(&self) -> &Note {
-        &self.note
-    }
-    fn memo(&self) -> Option<&MemoBytes> {
-        Some(&self.memo)
-    }
-    fn is_change(&self) -> bool {
-        self.transfer_type == TransferType::WalletInternal
-    }
-    fn nullifier(&self) -> Option<&Nullifier> {
-        None
     }
 }
 
@@ -693,107 +636,6 @@ pub(crate) fn truncate_to_height<P: consensus::Parameters>(
     Ok(())
 }
 
-/// Returns the commitment tree for the block at the specified height,
-/// if any.
-pub(crate) fn get_sapling_commitment_tree<P>(
-    wdb: &WalletDb<P>,
-    block_height: BlockHeight,
-) -> Result<Option<sapling::CommitmentTree>, SqliteClientError> {
-    wdb.conn
-        .query_row_and_then(
-            "SELECT sapling_tree FROM blocks WHERE height = ?",
-            [u32::from(block_height)],
-            |row| {
-                let row_data: Vec<u8> = row.get(0)?;
-                read_commitment_tree(&row_data[..]).map_err(|e| {
-                    rusqlite::Error::FromSqlConversionFailure(
-                        row_data.len(),
-                        rusqlite::types::Type::Blob,
-                        Box::new(e),
-                    )
-                })
-            },
-        )
-        .optional()
-        .map_err(SqliteClientError::from)
-}
-
-/// Returns the incremental witnesses for the block at the specified height,
-/// if any.
-pub(crate) fn get_sapling_witnesses<P>(
-    wdb: &WalletDb<P>,
-    block_height: BlockHeight,
-) -> Result<Vec<(NoteId, sapling::IncrementalWitness)>, SqliteClientError> {
-    let mut stmt_fetch_witnesses = wdb
-        .conn
-        .prepare("SELECT note, witness FROM sapling_witnesses WHERE block = ?")?;
-    let witnesses = stmt_fetch_witnesses
-        .query_map([u32::from(block_height)], |row| {
-            let id_note = NoteId::ReceivedNoteId(row.get(0)?);
-            let wdb: Vec<u8> = row.get(1)?;
-            Ok(read_incremental_witness(&wdb[..]).map(|witness| (id_note, witness)))
-        })
-        .map_err(SqliteClientError::from)?;
-
-    // unwrap database error & IO error from IncrementalWitness::read
-    let res: Vec<_> = witnesses.collect::<Result<Result<_, _>, _>>()??;
-    Ok(res)
-}
-
-/// Retrieves the set of nullifiers for "potentially spendable" Sapling notes that the
-/// wallet is tracking.
-///
-/// "Potentially spendable" means:
-/// - The transaction in which the note was created has been observed as mined.
-/// - No transaction in which the note's nullifier appears has been observed as mined.
-pub(crate) fn get_sapling_nullifiers<P>(
-    wdb: &WalletDb<P>,
-) -> Result<Vec<(AccountId, Nullifier)>, SqliteClientError> {
-    // Get the nullifiers for the notes we are tracking
-    let mut stmt_fetch_nullifiers = wdb.conn.prepare(
-        "SELECT rn.id_note, rn.account, rn.nf, tx.block as block
-         FROM sapling_received_notes rn
-         LEFT OUTER JOIN transactions tx
-         ON tx.id_tx = rn.spent
-         WHERE block IS NULL
-         AND nf IS NOT NULL",
-    )?;
-    let nullifiers = stmt_fetch_nullifiers.query_map([], |row| {
-        let account: u32 = row.get(1)?;
-        let nf_bytes: Vec<u8> = row.get(2)?;
-        Ok((
-            AccountId::from(account),
-            Nullifier::from_slice(&nf_bytes).unwrap(),
-        ))
-    })?;
-
-    let res: Vec<_> = nullifiers.collect::<Result<_, _>>()?;
-    Ok(res)
-}
-
-/// Returns the nullifiers for the notes that this wallet is tracking.
-pub(crate) fn get_all_sapling_nullifiers<P>(
-    wdb: &WalletDb<P>,
-) -> Result<Vec<(AccountId, Nullifier)>, SqliteClientError> {
-    // Get the nullifiers for the notes we are tracking
-    let mut stmt_fetch_nullifiers = wdb.conn.prepare(
-        "SELECT rn.id_note, rn.account, rn.nf
-         FROM sapling_received_notes rn
-         WHERE nf IS NOT NULL",
-    )?;
-    let nullifiers = stmt_fetch_nullifiers.query_map([], |row| {
-        let account: u32 = row.get(1)?;
-        let nf_bytes: Vec<u8> = row.get(2)?;
-        Ok((
-            AccountId::from(account),
-            Nullifier::from_slice(&nf_bytes).unwrap(),
-        ))
-    })?;
-
-    let res: Vec<_> = nullifiers.collect::<Result<_, _>>()?;
-    Ok(res)
-}
-
 /// Returns unspent transparent outputs that have been received by this wallet at the given
 /// transparent address, such that the block that included the transaction was mined at a
 /// height less than or equal to the provided `max_height`.
@@ -895,7 +737,7 @@ pub(crate) fn insert_block<'a, P>(
     block_height: BlockHeight,
     block_hash: BlockHash,
     block_time: u32,
-    commitment_tree: &sapling::CommitmentTree,
+    commitment_tree: &CommitmentTree,
 ) -> Result<(), SqliteClientError> {
     stmts.stmt_insert_block(block_height, block_hash, block_time, commitment_tree)
 }
@@ -935,20 +777,6 @@ pub(crate) fn put_tx_data<'a, P>(
         // It was there, so grab its row number.
         stmts.stmt_select_tx_ref(&txid)
     }
-}
-
-/// Marks a given nullifier as having been revealed in the construction
-/// of the specified transaction.
-///
-/// Marking a note spent in this fashion does NOT imply that the
-/// spending transaction has been mined.
-pub(crate) fn mark_sapling_note_spent<'a, P>(
-    stmts: &mut DataConnStmtCache<'a, P>,
-    tx_ref: i64,
-    nf: &Nullifier,
-) -> Result<(), SqliteClientError> {
-    stmts.stmt_mark_sapling_note_spent(tx_ref, nf)?;
-    Ok(())
 }
 
 /// Marks the given UTXO as having been spent.
@@ -1003,67 +831,6 @@ pub(crate) fn put_received_transparent_utxo<'a, P: consensus::Parameters>(
                     }
                 })
         })
-}
-
-/// Records the specified shielded output as having been received.
-///
-/// This implementation relies on the facts that:
-/// - A transaction will not contain more than 2^63 shielded outputs.
-/// - A note value will never exceed 2^63 zatoshis.
-pub(crate) fn put_received_note<'a, P, T: ReceivedSaplingOutput>(
-    stmts: &mut DataConnStmtCache<'a, P>,
-    output: &T,
-    tx_ref: i64,
-) -> Result<NoteId, SqliteClientError> {
-    let rcm = output.note().rcm().to_repr();
-    let account = output.account();
-    let to = output.note().recipient();
-    let diversifier = to.diversifier();
-    let value = output.note().value();
-    let memo = output.memo();
-    let is_change = output.is_change();
-    let output_index = output.index();
-    let nf = output.nullifier();
-
-    // First try updating an existing received note into the database.
-    if !stmts.stmt_update_received_note(
-        account,
-        diversifier,
-        value.inner(),
-        rcm,
-        nf,
-        memo,
-        is_change,
-        tx_ref,
-        output_index,
-    )? {
-        // It isn't there, so insert our note into the database.
-        stmts.stmt_insert_received_note(
-            tx_ref,
-            output_index,
-            account,
-            diversifier,
-            value.inner(),
-            rcm,
-            nf,
-            memo,
-            is_change,
-        )
-    } else {
-        // It was there, so grab its row number.
-        stmts.stmt_select_received_note(tx_ref, output.index())
-    }
-}
-
-/// Records the incremental witness for the specified note,
-/// as of the given block height.
-pub(crate) fn insert_witness<'a, P>(
-    stmts: &mut DataConnStmtCache<'a, P>,
-    note_id: i64,
-    witness: &sapling::IncrementalWitness,
-    height: BlockHeight,
-) -> Result<(), SqliteClientError> {
-    stmts.stmt_insert_witness(NoteId::ReceivedNoteId(note_id), height, witness)
 }
 
 /// Removes old incremental witnesses up to the given block height.
