+++ zcash_extensions/Cargo.toml
@@ -11,14 +11,14 @@ rust-version = "1.65"
 
 [dependencies]
 blake2b_simd = "1"
-zcash_primitives = { version = "0.11", path = "../zcash_primitives", default-features = false, features = ["zfuture" ] }
+zcash_primitives = { version = "0.12", path = "../zcash_primitives", default-features = false, features = ["zfuture" ] }
 
 [dev-dependencies]
 ff = "0.13"
 jubjub = "0.10"
 rand_core = "0.6"
-zcash_address = { version = "0.2", path = "../components/zcash_address" }
-zcash_proofs = { version = "0.11", path = "../zcash_proofs" }
+zcash_address = { version = "0.3", path = "../components/zcash_address" }
+zcash_proofs = { version = "0.12", path = "../zcash_proofs" }
 
 [features]
 transparent-inputs = []
