+++ zcash_proofs/CHANGELOG.md
@@ -6,7 +6,10 @@ and this library adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.12.0] - 2023-06-06
 ### Changed
+- Bumped dependencies to `incrementalmerkletree 0.4`, `zcash_primitives 0.12`
 - MSRV is now 1.65.0.
 
 ### Removed
